more of testing done completed msot achieved 40 eprcent coverage
