--- ToDoList/.idea/data_source_mapping.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="DataSourcePerFileMappings">
-    <file url="file://$APPLICATION_CONFIG_DIR$/consoles/db/d62886af-db77-46e2-86f1-f1991c2f8043/console.sql" value="d62886af-db77-46e2-86f1-f1991c2f8043" />
-  </component>
-</project>
