+++ ToDoList/demo/src/main/java/com/todolist/Models/updateTaskRequestPackage.java
@@ -1,0 +1,4 @@
+package com.todolist.Models;
+
+public class updateTaskRequestPackage {
+}
