--- ToDoList/demo/src/main/java/com/todolist/Services/addTaskService.java
@@ -1,4 +1,0 @@
-package com.todolist.Services;
-
-public class addTaskService {
-}
