--- ToDoList/demo/src/main/java/com/todolist/Services/taskRetrivalService.java
@@ -1,4 +1,0 @@
-package com.todolist.Services;
-
-public class taskRetrivalService {
-}
