+++ ToDoList/demo/src/test/java/com/todolist/adaptors/web/AdaptorServiceTest.java
@@ -1,0 +1,4 @@
+import static org.junit.jupiter.api.Assertions.*;
+class AdaptorServiceTest {
+  
+}
